pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7.3"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
peewee = "^3.15.4"
python-dotenv = "^0.21.1"
bcrypt = "^4.0.1"
PyJWT = "^2.6.0"
click = "^8.1.3"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "src.interfaces.cli.cli:cli"
src = "src.app:main"